backend/apps/material/serializers.py
from rest_framework import serializers
from apps.material.models import Color


class ColorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Color
        fields = ('name', 'hex',)