backend/requirements.txt
asgiref==3.4.1
Django~=4.0
django-cors-headers==3.10.0
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
zipp==3.6.0
Unidecode
django-colorfield==0.7.2
Faker
requests
django-jet
singlemodeladmin==0.9
colorthief==0.2.1
numpy==1.23.5
django-admin-inline-paginator==0.3.0
drf-writable-nested==0.7.0
